Updated the radio button layout
