+++ pages/vote/radioLayout.vue
@@ -1,0 +1,26 @@
+<template>
+  <div>
+    <div
+      v-for="childTitle in titles"
+      :key="childTitle"
+      >
+      {{ childTitle }}
+      <div
+        v-for="child in children"
+        :key="child.title"
+      >
+      <input type="radio" :name="childTitle"/>
+      <label>{{ child }}</label>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  props: {
+    children: Array,
+    titles: Array
+  }
+}
+</script>
